Share IT test bundle
